updated makefile to show instructions and added compiler flags. removed unused vars from srcs removed add_to_heap from QP. added Wextra compiler flag added config.h as required header. created a config script removed std includes and including them using config.h

+++ MinHeap.cpp
@@ -1,4 +1,4 @@
-#include <iostream>
+#include "config.h"
 #include "MinHeap.h"
 using namespace std;
 
@@ -9,7 +9,7 @@ MinHeap::MinHeap(int size)
 
 void MinHeap::push(Score score)
 {
-    if( heap.size() < max_size )
+    if( ( int )heap.size() < max_size )
         heap.push( score );
     else {
         if( heap.top().get_score() < score.get_score() ) {
